关闭分享bug修正

+++ mbstyle.go
@@ -60,12 +60,12 @@ type Root struct {
 	// Default bearing, in degrees clockwise from true north. The style bearing
 	// will be used only if the map has not been positioned by other means
 	// (e.g. map options or user interaction).
-	Bearing *int64 `json:"bearing,omitempty"`
+	Bearing *float64 `json:"bearing,omitempty"`
 	// Default pitch, in degrees. Zero is perpendicular to the surface, for a
 	// look straight down at the map, while a greater value like 60 looks ahead
 	// towards the horizon. The style pitch will be used only if the map has not
 	// been positioned by other means (e.g. map options or user interaction).
-	Pitch *int64 `json:"pitch,omitempty"`
+	Pitch *float64 `json:"pitch,omitempty"`
 	// The global light source.
 	Light *Light `json:"light,omitempty"`
 	// Data source specifications.
